// frontend/src/api/models/Person.ts
export interface Person {
    /**
     * 
     * @type {number}
     * @memberof Person
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Person
     */
    name: string;
    /**
     * 
     * @type {Coordinates}
     * @memberof Person
     */
    coordinates: Coordinates;
    /**
     * 
     * @type {Date}
     * @memberof Person
     */
    creationDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof Person
     */
    eyeColor: PersonEyeColorEnum;
    /**
     * 
     * @type {string}
     * @memberof Person
     */
    hairColor?: PersonHairColorEnum;
    /**
     * 
     * @type {Location}
     * @memberof Person
     */
    location: Location;
    /**
     * 
     * @type {number}
     * @memberof Person
     */
    height: number;
    /**
     * 
     * @type {string}
     * @memberof Person
     */
    nationality?: PersonNationalityEnum;
}


/**
 * @export
 */
export const PersonEyeColorEnum = {
    Green: 'GREEN',
    Orange: 'ORANGE',
    White: 'WHITE',
    Brown: 'BROWN'
} as const;
export type PersonEyeColorEnum = typeof PersonEyeColorEnum[keyof typeof PersonEyeColorEnum];

/**
 * @export
 */
export const PersonHairColorEnum = {
    Green: 'GREEN',
    Orange: 'ORANGE',
    White: 'WHITE',
    Brown: 'BROWN'
} as const;
export type PersonHairColorEnum = typeof PersonHairColorEnum[keyof typeof PersonHairColorEnum];

/**
 * @export
 */
export const PersonNationalityEnum = {
    UnitedKingdom: 'UNITED_KINGDOM',
    India: 'INDIA',
    Italy: 'ITALY',
    NorthKorea: 'NORTH_KOREA'
} as const;
export type PersonNationalityEnum = typeof PersonNationalityEnum[keyof typeof PersonNationalityEnum];